+++ samples/SampleWebApplication/Program.cs
@@ -1,0 +1,106 @@
+using Azure.Data.Tables;
+
+using Microsoft.Extensions.DependencyInjection.Extensions;
+
+using Serilog;
+using Serilog.Events;
+
+namespace SampleWebApplication;
+
+public class Program
+{
+    private const string OutputTemplate = "{Timestamp:HH:mm:ss.fff} [{Level:u1}] {Message:lj}{NewLine}{Exception}";
+
+    public static async Task<int> Main(string[] args)
+    {
+        // azure home directory
+        var homeDirectory = Environment.GetEnvironmentVariable("HOME") ?? ".";
+        var logDirectory = Path.Combine(homeDirectory, "LogFiles");
+
+        Log.Logger = new LoggerConfiguration()
+            .MinimumLevel.Verbose()
+            .MinimumLevel.Override("Microsoft", LogEventLevel.Information)
+            .Enrich.FromLogContext()
+            .WriteTo.Console(outputTemplate: OutputTemplate)
+            .CreateBootstrapLogger();
+
+        try
+        {
+            Log.Information("Starting web host");
+
+            var builder = WebApplication.CreateBuilder(args);
+
+            builder.Host
+                .UseSerilog((context, services, configuration) => configuration
+                    .ReadFrom.Configuration(context.Configuration)
+                    .ReadFrom.Services(services)
+                    .Enrich.FromLogContext()
+                    .Enrich.WithProperty("ApplicationName", builder.Environment.ApplicationName)
+                    .Enrich.WithProperty("EnvironmentName", builder.Environment.EnvironmentName)
+                    .WriteTo.Console(outputTemplate: OutputTemplate)
+                    .WriteTo.File(
+                        path: $"{logDirectory}/log.txt",
+                        rollingInterval: RollingInterval.Day,
+                        shared: true,
+                        flushToDiskInterval: TimeSpan.FromSeconds(1),
+                        outputTemplate: OutputTemplate,
+                        retainedFileCountLimit: 10
+                    )
+                    .WriteTo.AzureTableStorage(
+                        storageAccount: services.GetRequiredService<TableServiceClient>(),
+                        storageTableName: "SampleLog",
+                        propertyColumns: new[] { "SourceContext", "RequestId", "RequestPath", "ConnectionId", "ApplicationName", "EnvironmentName" }
+                    )
+                );
+
+            ConfigureServices(builder);
+
+            var app = builder.Build();
+
+            ConfigureMiddleware(app);
+
+            await app.RunAsync();
+
+            return 0;
+        }
+        catch (Exception ex)
+        {
+            Log.Fatal(ex, "Host terminated unexpectedly");
+            return 1;
+        }
+        finally
+        {
+            await Log.CloseAndFlushAsync();
+        }
+    }
+
+    private static void ConfigureServices(WebApplicationBuilder builder)
+    {
+        builder.Services.AddRazorPages();
+
+        builder.Services.TryAddSingleton<TableServiceClient>(sp =>
+        {
+            var configuration = sp.GetRequiredService<IConfiguration>();
+            var connectionString = configuration.GetConnectionString("StorageAccount");
+            return new TableServiceClient(connectionString);
+        });
+    }
+
+    private static void ConfigureMiddleware(WebApplication app)
+    {
+        if (!app.Environment.IsDevelopment())
+        {
+            app.UseExceptionHandler("/Error");
+            app.UseHsts();
+        }
+
+        app.UseHttpsRedirection();
+        app.UseStaticFiles();
+
+        app.UseRouting();
+
+        app.UseAuthorization();
+
+        app.MapRazorPages();
+    }
+}